feat: add commands to BlockSpec and register commands in spec store (#7898)
### What changed?
- Add commands in `BlockSpec`, which can be easily expanded externally
- Register declared commands in `spec-store`
- Refactor `embed-linked-doc-block` and remove useless service
