+++ packages/blocks/src/_common/configs/quick-action/config.ts
@@ -143,15 +143,7 @@ export const quickActionConfig: QuickActionConfig[] = [
       const autofill = getTitleFromSelectedModels(selectedModels);
       void promptDocTitle(host, autofill).then(title => {
         if (title === null) return;
-        const linkedDoc = convertSelectedBlocksToLinkedDoc(
-          doc,
-          selectedModels,
-          title
-        );
-        const linkedDocService = host.spec.getService(
-          'affine:embed-linked-doc'
-        );
-        linkedDocService.slots.linkedDocCreated.emit({ docId: linkedDoc.id });
+        convertSelectedBlocksToLinkedDoc(doc, selectedModels, title);
         notifyDocCreated(host, doc);
       });
     },
