+++ packages/blocks/src/embed-linked-doc-block/commands/index.ts
@@ -1,0 +1,7 @@
+import type { BlockCommands } from '@blocksuite/block-std';
+
+import { insertLinkByQuickSearchCommand } from './insert-link-by-quick-search.js';
+
+export const commands: BlockCommands = {
+  insertLinkByQuickSearch: insertLinkByQuickSearchCommand,
+};
