+++ packages/blocks/src/embed-linked-doc-block/embed-linked-doc-block.ts
@@ -1,5 +1,3 @@
-import type { EditorHost } from '@blocksuite/block-std';
-
 import { Bound } from '@blocksuite/global/utils';
 import { assertExists } from '@blocksuite/global/utils';
 import { DocCollection } from '@blocksuite/store';
@@ -11,6 +9,7 @@ import type { DocMode } from '../_common/types.js';
 import type { RootBlockComponent } from '../root-block/index.js';
 import type { SurfaceRefBlockService } from '../surface-ref-block/index.js';
 import type { SurfaceRefRenderer } from '../surface-ref-block/surface-ref-renderer.js';
+import type { EmbedLinkedDocBlockConfig } from './embed-linked-doc-config.js';
 import type {
   EmbedLinkedDocModel,
   EmbedLinkedDocStyles,
@@ -25,11 +24,6 @@ import { renderLinkedDocInCard } from '../_common/utils/render-linked-doc.js';
 import { SyncedDocErrorIcon } from '../embed-synced-doc-block/styles.js';
 import { styles } from './styles.js';
 import { getEmbedLinkedDocIcons } from './utils.js';
-
-export interface EmbedLinkedDocBlockConfig {
-  handleClick?: (e: MouseEvent, host: EditorHost) => void;
-  handleDoubleClick?: (e: MouseEvent, host: EditorHost) => void;
-}
 
 @customElement('affine-embed-linked-doc-block')
 @Peekable({
