+++ packages/blocks/src/embed-linked-doc-block/embed-linked-doc-model.ts
@@ -27,5 +27,8 @@ declare global {
     interface EdgelessBlockModelMap {
       'affine:embed-linked-doc': EmbedLinkedDocModel;
     }
+    interface BlockModels {
+      'affine:embed-linked-doc': EmbedLinkedDocModel;
+    }
   }
 }
