+++ packages/blocks/src/embed-linked-doc-block/embed-linked-doc-spec.ts
@@ -2,14 +2,13 @@ import type { BlockSpec } from '@blocksuite/block-std';
 
 import { literal } from 'lit/static-html.js';
 
-import type { EmbedLinkedDocBlockConfig } from './embed-linked-doc-block.js';
+import type { EmbedLinkedDocBlockConfig } from './embed-linked-doc-config.js';
 
+import { commands } from './commands/index.js';
 import { EmbedLinkedDocBlockSchema } from './embed-linked-doc-schema.js';
-import { EmbedLinkedDocBlockService } from './embed-linked-doc-service.js';
 
 export type EmbedLinkedDocBlockSpecType = BlockSpec<
   string,
-  EmbedLinkedDocBlockService,
   EmbedLinkedDocBlockConfig
 >;
 
@@ -18,5 +17,5 @@ export const EmbedLinkedDocBlockSpec: EmbedLinkedDocBlockSpecType = {
   view: {
     component: literal`affine-embed-linked-doc-block`,
   },
-  service: EmbedLinkedDocBlockService,
+  commands,
 };
