+++ packages/framework/block-std/src/spec/type.ts
@@ -2,6 +2,7 @@ import type { DisposableGroup } from '@blocksuite/global/utils';
 import type { BlockSchemaType } from '@blocksuite/store';
 import type { StaticValue } from 'lit/static-html.js';
 
+import type { Command } from '../command/types.js';
 import type { BlockService } from '../service/index.js';
 import type { BlockServiceConstructor } from '../service/index.js';
 import type { BlockSpecSlots } from './slots.js';
@@ -11,14 +12,17 @@ export interface BlockView<WidgetNames extends string = string> {
   widgets?: Record<WidgetNames, StaticValue>;
 }
 
+export type BlockCommands = Partial<Record<keyof BlockSuite.Commands, Command>>;
+
 export interface BlockSpec<
   WidgetNames extends string = string,
-  Service extends BlockService = BlockService,
   BlockConfig = object,
+  Service extends BlockService = BlockService,
 > {
   schema: BlockSchemaType;
   view: BlockView<WidgetNames>;
   config?: BlockConfig;
+  commands?: BlockCommands;
   service?: BlockServiceConstructor<Service>;
   setup?: (slots: BlockSpecSlots, disposableGroup: DisposableGroup) => void;
 }
