+++ packages/playground/apps/starter/utils/editor.ts
@@ -56,17 +56,13 @@ export async function mountDefaultDocEditor(collection: DocCollection) {
   const editor = new AffineEditorContainer();
   editor.pageSpecs = [...PageEditorBlockSpecs].map(spec => {
     if (spec.schema.model.flavour === 'affine:page') {
-      spec = patchPageRootSpec(
-        spec as BlockSpec<'affine:page', PageRootService>
-      );
+      spec = patchPageRootSpec(spec);
     }
     return spec;
   });
   editor.edgelessSpecs = [...EdgelessEditorBlockSpecs].map(spec => {
     if (spec.schema.model.flavour === 'affine:page') {
-      spec = patchPageRootSpec(
-        spec as BlockSpec<'affine:page', PageRootService>
-      );
+      spec = patchPageRootSpec(spec);
     }
     return spec;
   });
@@ -148,7 +144,7 @@ export async function mountDefaultDocEditor(collection: DocCollection) {
 
   return editor;
 
-  function patchPageRootSpec(spec: BlockSpec<'affine:page', PageRootService>) {
+  function patchPageRootSpec(spec: BlockSpec) {
     const setup = spec.setup;
     const newSpec: typeof spec = {
       ...spec,
